did the delete endpoint and got it working

+++ src/main/java/com/journal_app/backend_journal_app/repository/EntriesRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.UUID;
 
 public interface EntriesRepository
-        extends JpaRepository<Entries, Integer> {
+        extends JpaRepository<Entries, Long> {
 
     List<Entries> findByUsers(UUID users);
